user authorization

+++ Back-End/controllers/feed.js
@@ -2,6 +2,7 @@ const {validationResult} = require('express-validator/check');
 const Post = require('../models/post');
 const fs = require('fs');
 const path = require('path');
+const User = require('../models/user');
 
 
 exports.getPosts = async (req, res, next) => {
@@ -21,9 +22,10 @@ exports.getPosts = async (req, res, next) => {
 };
 
 exports.createPost = (req, res, next) => {
-    console.log(req);
+    // console.log(req);
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     console.log(title);
 
     const errors = validationResult(req);
@@ -41,21 +43,28 @@ exports.createPost = (req, res, next) => {
     }
     const imageUrl = req.file.path.replace("\\" ,"/");
 
-
     
     const post = new Post({
         title: title, 
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: "John Doe"
-        }
+        creator: req.userId
     });
-    post.save().then(result => {
-        console.log(result);
+    post
+    .save()
+    .then(result => {
+        return User.findById(req.userId);
+    })
+    .then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    })
+    .then(result => {
         res.status(201).json({
             message: "Post created successfully",
-            post: result
+            post: post,
+            creator: {_id: creator._id, name: creator.name}
         });
     })
     .catch(err => {
@@ -118,6 +127,11 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+        if (post.creator.toString() !== req.userId){
+            const error = new Error('Not Authorized');
+            error.statusCode = 403; // authorization issue
+            throw error
+        }
         if (imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl);
         }
@@ -148,7 +162,11 @@ exports.deletePost = (req, res, next) => {
             throw error;
         }
         // check logged in user
-
+        if (post.creator.toString() !== req.userId){
+            const error = new Error('Not Authorized');
+            error.statusCode = 403; // authorization issue
+            throw error
+        }
         clearImage(post.imageUrl);
         return post.deleteOne();
     })
